# config/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Datix API",
    "version": "0.1.0",
    "description": "Cookie-authenticated website management and reports; public, origin-validated event collection. Reports are eventually consistent with queue processing. Visitor counts are estimates scoped to an environment and UTC day."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/tracker-config": {
      "get": {
        "operationId": "getTrackerConfig",
        "description": "Public tracking switches only. The tracker refreshes these after 60 seconds. No account or billing data is returned.",
        "parameters": [
          {
            "name": "siteId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "environmentId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "enabled": {
                      "type": "boolean"
                    },
                    "settings": {
                      "type": "object",
                      "additionalProperties": false,
                      "description": "Per-environment switches. Omitted persisted keys default to true. When updating, send every switch as a boolean.",
                      "properties": {
                        "pageview": {
                          "type": "boolean"
                        },
                        "custom": {
                          "type": "boolean"
                        },
                        "click": {
                          "type": "boolean"
                        },
                        "outbound": {
                          "type": "boolean"
                        },
                        "download": {
                          "type": "boolean"
                        },
                        "form_submit": {
                          "type": "boolean"
                        },
                        "scroll": {
                          "type": "boolean"
                        },
                        "engagement": {
                          "type": "boolean"
                        },
                        "referrer": {
                          "type": "boolean"
                        },
                        "country": {
                          "type": "boolean"
                        },
                        "device": {
                          "type": "boolean"
                        },
                        "dimensions": {
                          "type": "boolean"
                        },
                        "language": {
                          "type": "boolean"
                        },
                        "coordinates": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "400 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "401 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "403 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "404 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "409 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "413 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "415 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "429 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "503 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/usage": {
      "get": {
        "operationId": "getAccountUsage",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SecureSessionCookie": []
          }
        ],
        "description": "Account-wide event allowance, monthly renewal window and per-website collection status. Requires an active subscription or trial to collect. Production pageviews use 1 credit, actions use 0.5; localhost multiplies these by 0.3. Engagement heartbeats and duplicate deliveries are free.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "plan": {
                      "type": [
                        "object",
                        "null"
                      ],
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "trial": {
                          "type": "boolean"
                        },
                        "eventLimit": {
                          "type": "integer"
                        },
                        "websiteLimit": {
                          "type": "integer"
                        }
                      }
                    },
                    "period": {
                      "type": [
                        "object",
                        "null"
                      ],
                      "properties": {
                        "start": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "end": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "events": {
                      "type": "object",
                      "properties": {
                        "used": {
                          "type": "number",
                          "multipleOf": 0.01
                        },
                        "remaining": {
                          "type": "number",
                          "multipleOf": 0.01
                        }
                      }
                    },
                    "protection": {
                      "type": "object",
                      "properties": {
                        "since": {
                          "type": "string",
                          "format": "date"
                        },
                        "blocked": {
                          "type": "integer"
                        },
                        "lastBlockedAt": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "format": "date-time"
                        },
                        "learning": {
                          "type": "integer"
                        },
                        "learned": {
                          "type": "integer"
                        },
                        "reasons": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "reason": {
                                "enum": [
                                  "source_limit",
                                  "repeated_activity",
                                  "unusual_activity"
                                ]
                              },
                              "blocked": {
                                "type": "integer"
                              }
                            }
                          }
                        }
                      }
                    },
                    "paused": {
                      "type": "boolean"
                    },
                    "pauseReason": {
                      "enum": [
                        null,
                        "subscription_required",
                        "event_limit",
                        "website_limit",
                        "website_budget",
                        "disabled"
                      ]
                    },
                    "websites": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string"
                          },
                          "domain": {
                            "type": "string"
                          },
                          "events": {
                            "type": "number",
                            "multipleOf": 0.01
                          },
                          "creditBudget": {
                            "type": [
                              "number",
                              "null"
                            ],
                            "minimum": 0.15,
                            "maximum": 1000000000,
                            "multipleOf": 0.01
                          },
                          "paused": {
                            "type": "boolean"
                          },
                          "pauseReason": {
                            "enum": [
                              null,
                              "subscription_required",
                              "event_limit",
                              "website_limit",
                              "website_budget",
                              "disabled"
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "400 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "401 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "403 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "404 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "409 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "413 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "415 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "429 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "503 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me": {
      "get": {
        "operationId": "getCurrentUser",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SecureSessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "400 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "401 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "403 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "404 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "409 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "413 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "415 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "429 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "503 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sites/{siteId}/installation": {
      "parameters": [
        {
          "name": "siteId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "environment",
          "in": "query",
          "description": "Environment ID belonging to this site. Omit for its default Production environment.",
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getInstallationStatus",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SecureSessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "receiving": {
                      "type": "boolean",
                      "description": "At least one persisted pageview remains in raw retention."
                    },
                    "lastReceivedAt": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "400 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "401 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "403 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "404 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "409 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "413 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "415 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "429 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "503 error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/billing": {
      "get": {
        "operationId": "getBilling",
        "description": "Refresh Autumn customer availability.",
        "security": [
          {
            "SessionCookie": []
          },
          {
            "SecureSessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "hasCustomer": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Billing unavailable or unconfigured"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      },
      "SecureSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-better-auth.session_token"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        }
      },
...
    }
  }
}